Changing storage of CCModel to store VID and EID instead of CfgModel and EdgeModel copies.
git-svn-id: svn+ssh://parasol-svn.cs.tamu.edu/research/parasol-svn/svnrepository/vizmo/trunk@521 8e255c04-9930-0410-9bf1-9572f6c5958d

+++ src/Models/CfgModel.cpp
@@ -42,7 +42,7 @@ CfgModel::SetCfg(const vector<double>& _newCfg) {
 }
 
 void
-CfgModel::Set(int _index, CCModel<CfgModel, EdgeModel>* _cc){
+CfgModel::Set(size_t _index, CCModel<CfgModel, EdgeModel>* _cc){
   m_index = _index;
   SetName();
   m_cc = _cc;
